+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 > 2016-02-18
 
 * Save actions with meta.cookie to browser cookie. Added `cookies-js` dependency. If you don't want it please make a `redux-cookie` module specific to that use case.
+* Pass cookieJar as `cookie` meta to server.